ISSUE: 99; ISSUE: 112; ISSUE: 140; removed CSRF decorator in users.views

+++ enrollment/records/models.py
@@ -191,10 +191,16 @@ class Record( models.Model ):
             student = user.student  
             old_group = Group.objects.get(id=old_id)
             new_group = Group.objects.get(id=new_id)
-            record = Record.enrolled.get(group=old_group, student=student)
-            record.delete()
-            new_record = Record.objects.create(group=new_group, student=student, status=STATUS_ENROLLED)
-            return new_record
+            student_options = StudentOptions.get_student_options_for_subject(student.id, new_group.subject.id)
+            if not student_options.is_recording_open():
+                raise RecordsNotOpenException()
+            if Record.number_of_students(group=new_group) < new_group.limit:
+                record = Record.enrolled.get(group=old_group, student=student)
+                record.delete()
+                new_record = Record.objects.create(group=new_group, student=student, status=STATUS_ENROLLED)
+            else:
+                raise OutOfLimitException()
+            return new_record   
         except Student.DoesNotExist:
             raise NonStudentException()
         except Group.DoesNotExist:
